Add documentation for using GitHub App Installation tokens instead of personal access tokens.
Identified this while using an app installation token locally and working through the code.
It seems that the use case for GitHub App Installations is currently only for specifically
running from within GitHub Actions, which sets GITHUB_TOKEN rather than GH_TOKEN. However,
it is possible that other CI tools make use of GitHub App Installation tokens, and thus they
also require the same treatment.

This could be fixed in the code to be more explicit, but for now hopefully the documentation
change is helpful to some.